se implemeto diseño

+++ src/app/app.component.ts
@@ -1,5 +1,7 @@
 import { Component } from '@angular/core';
 import { LoginService } from './service/login.service';
+import { MatDialog } from '@angular/material/dialog';
+import { CerrarSesionComponent } from './component/dialogo/cerrar-sesion/cerrar-sesion.component';
 
 @Component({
   selector: 'app-root',
@@ -10,11 +12,21 @@ export class AppComponent {
   title = 'WoodWorkSmartAp';
 
   role:string="";
-  constructor(private loginService: LoginService) {
+  constructor(private loginService: LoginService, private dialog: MatDialog) {
   }
 
   cerrar() {
-    sessionStorage.clear();
+    const dialogRef = this.dialog.open(CerrarSesionComponent, {
+      width: '300px',
+      data: { mensaje: '¿Estás seguro de cerrar la sesión?' }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        sessionStorage.clear();
+        window.location.href = '/login';
+      }
+    });
   }
 
   verificar() {
@@ -28,5 +40,7 @@ export class AppComponent {
       return false;
     }
   }
+
+
 
 }
